% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_int_grfx.R
\name{agg_kw}
\alias{agg_kw}
\title{Roll Up kW}
\usage{
agg_kw(x, .fn = mean)
}
\arguments{
\item{x}{Grouped data with kwh - note \code{"kW"} should be a column in the
data.}

\item{.fn}{Aggregating function, default is \code{base::mean()}}

\item{...}{Additional arguments to pass to .fn}
}
\description{
Aggregate kW from session or interval, etc. data to a specified
level - pass data grouped to the desired level.
}
