% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_l2_gen_sess.R
\name{agg_kwh}
\alias{agg_kwh}
\title{Roll Up kWh}
\usage{
agg_kwh(x, .fn = sum, ...)
}
\arguments{
\item{x}{Grouped data with kwh - note \code{"kwh"} should be a column in the
data.}

\item{.fn}{Aggregating function, default is \code{base::sum()}}

\item{...}{Additional arguments to pass to .fn}
}
\description{
Aggregate kwh from session or interval, etc. data to a specified
level - pass data grouped to the desired level.
}
