% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_afdc_data.R
\name{checkRenv}
\alias{checkRenv}
\title{Check .Renviron gitignore}
\usage{
checkRenv()
}
\description{
Check if \code{.Renviron} file is in \code{.gitignore}. If it
isn't, prompts user to add to \code{.gitignore} file not make sure api keys/other
secrets don't be exposed.
}
\examples{
\dontrun{
checkRenv()
}
}
