% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_l2_gen_sess.R
\name{gen_l2_sess}
\alias{gen_l2_sess}
\title{Generate a level 2 charging Session for home chargers}
\usage{
gen_l2_sess(start, duration)
}
\arguments{
\item{start}{A datetime object, when the session starts}

\item{duration}{A double, number of hours the session lasts}
}
\description{
Generate a synthetic level 2 EVSE charging session
for a home charger. Resulting session comes in 15 minute intervals.
}
