% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_l2_gen_sess.R
\name{generate.interval}
\alias{generate.interval}
\alias{generate.interval.xgb.Booster}
\title{Generate 15 min intervals}
\usage{
generate.interval(start, duration, model)

\method{generate.interval}{xgb.Booster}(start, duration, model)
}
\arguments{
\item{start}{A datetime object, when the session starts}

\item{duration}{A double, number of hours the session lasts}

\item{model}{Model to use to generate intervals}
}
\description{
Generate 15 minute intervals for a level 2 charger
}
