% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_afdc_data.R
\name{getNrelKey}
\alias{getNrelKey}
\title{NREL API}
\usage{
getNrelKey()
}
\description{
Returns user's NREL API key. If an NREL key is not found,
prompts user to add it to session variables or to create a key from
\link{https://developer.nrel.gov/signup/} and then add to session variables.
}
\examples{
\dontrun{
getNrelKey()
}
}
