% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_int_grfx.R
\name{plt_chrg_quant}
\alias{plt_chrg_quant}
\title{Plot Load Shape Quantiles}
\usage{
plt_chrg_quant(.data, time_col, grp_col, col = kwh, .fn = sum)
}
\arguments{
\item{.data}{Dataframe of interval charging data}

\item{time_col}{Datetime column in .data}

\item{grp_col}{Column to group and determine quantiles from}

\item{col}{Column in .data to aggregate and produce a load profile for,
default is \code{kwh}}

\item{.fn}{Funciton to use to aggregate \code{col}, default is \code{sum}}
}
\description{
Plot Load Shape Quantiles
}
