% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_int_grfx.R
\name{plt_load}
\alias{plt_load}
\title{Plot Load}
\usage{
plt_load(
  .data,
  time_col,
  load_col,
  group_col = NULL,
  agg_fn = mean,
  time_agg = lubridate::hour,
  plt_cf = FALSE,
  cf_fn = utilityR::cf.int,
  peak_hrs = NULL,
  ...
)
}
\arguments{
\item{.data}{Charging interval data}

\item{time_col}{Interval time column in \code{.data}}

\item{load_col}{Column in \code{.data} with charging load}

\item{group_col}{Default is \code{NULL}. If passed, what column to break
load profiles out by, e.g. if a charging customer id column is passed, the output
will have a load profile for each customer included.}

\item{agg_fn}{Function to use to aggregate load data, default is \code{mean}}

\item{time_agg}{What time interval to aggregate to, default is \code{lubridate::hour}}

\item{plt_cf}{\code{TRUE}/\code{FALSE} - whether to plot confidence interval}

\item{cf_fn}{Function to calculate a confidence interval around the aggregation statistic. Default is \code{utilityR::cf.int}}

\item{...}{Additional arguments to pass to \code{plt_load}}
}
\description{
Plot an aggregated load profile.
}
\examples{
\dontrun{

plt_load(l2_custs, time_col = int.start, load_col = interval.kwh / 0.25, group_col = cust)

}
}
