% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_int_grfx.R
\name{plt_pk_hrs}
\alias{plt_pk_hrs}
\title{ggplot2 Utility peak hour shading}
\usage{
plt_pk_hrs(start, end, fill = "gray", alpha = 0.2)
}
\arguments{
\item{start}{Numeric for the starting hour of peak hours (e.g. \code{12})}

\item{end}{Numeric for the ending hour, e.g. if peak hours are 12pm - 8pm
start would be \code{12} and end would be \code{20}}

\item{alpha}{The alpha to shade with, default is \code{0.2}}

\item{color}{Color to shade with, default is \code{"gray"}}
}
\description{
Add shading for peak hours to your load profile plots
}
\examples{
\dontrun{
plt_load(l2_custs[l2_custs$cust == 1, ],
time_col = int.start,
load_col = interval.kwh / 0.25,
cf_fn = function(x) 0) +
  labs(x = "Time of Day (Hours)",
       y = "Average kW",
       title = "Load Profile for a Level 2 Charger") +
  plt_pk_hrs(start = 12, end = 20)
}
}
