% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_afdc_data.R
\name{queryNrel}
\alias{queryNrel}
\title{NREL API}
\usage{
queryNrel(dataset = NULL, ...)
}
\arguments{
\item{dataset}{The Nrel dataset to return, call \code{getNrelArgs()} for a
list of datasets. For help setting up API credentials call \code{queryNrel()}}
}
\description{
Query the NREL vehicle API. Returns the dataset requested from
NREL (for a list of possible dat sets call \code{getNrelArgs()}). Requires
an NREL API key to use, calling \code{queryNrel()} will help walk you through
setting up the right credentials.
}
\examples{
\dontrun{
# Get NREL fuels
queryNrel("fuels")

# Get all vehicles
queryNrel("vehicles")

# Set up credentials
queryNrel()
}
}
