% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_afdc_data.R
\name{removeNrelKey}
\alias{removeNrelKey}
\title{NREL_API}
\usage{
removeNrelKey(perm = FALSE)
}
\arguments{
\item{perm}{If \code{TRUE} removes your NREL API key from your \code{.Renviron} file. Default is \code{FALSE}}
}
\description{
Remove your NREL API key from system variables
}
\examples{
\dontrun{
# Remove from session only
removeNrelKey()

# Remove from `.Renviron`
removeNrelKey(perm = TRUE)
}
}
