% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_session_expansion.R
\name{sess2int}
\alias{sess2int}
\title{Sessions to Intervals}
\usage{
sess2int(
  .data,
  ...,
  session_id = session_id,
  plug_in = plug_in,
  plug_out = plug_out,
  charge_start = charge_start,
  charge_stop = charge_stop,
  kwh = kwh,
  units = "mins"
)
}
\arguments{
\item{.data}{Session data to convert to intervals}

\item{...}{Any unquoted id variables. Examples would be
\code{port_id, charger_id, driver_id}.
@param plug_in The plug in time of the session, should be a datetime column.
@param plug_out The plug out time of the session, should be a datetime column.
@param charge_start The charge start time of the session, should be a datetime column.
@param charge_stop The charge stop time of the session, should be a datetime column.
@param kwh A numeric column representing the energy consumed during the
charging session in kWh.
@param units What kind of interval data to return. The default is minute
interval data, setting to \code{"15 minutes"} returns 15 minute intervals.
@example \dontrun{
 sess2int(
          session_data1,
          cust_id, charger_id, port_id = charger_id,
          units = "15 minutes"
              )
 }}
}
\description{
Convert PEV charging sessions to interval data.
Preferably a session data set with 1 row per charging session.
}
