% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_l2_gen_sess.R
\name{syn_l2_custs}
\alias{syn_l2_custs}
\title{Generate EV Owners' Level 2 Interval Data}
\usage{
syn_l2_custs(start, end, num_customers = 10, seed = 123)
}
\arguments{
\item{start}{A datetime object, minimum time owners could start charging}

\item{end}{A datetime object, upper bound for when owners can charge. Charge
sessions will occur between \code{start} and \code{end}.}

\item{num_customers}{Numeric for how many electric vehicle owners to include.}

\item{seed}{Random seed}
}
\description{
Generate Level 2 EVSE home charging data for multiple EV owners.
}
