% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_int_grfx.R
\name{plt_load}
\alias{plt_load}
\title{Plot Load}
\usage{
plt_load(
  .data,
  time_col,
  load_col,
  agg_fn = mean,
  time_agg = lubridate::hour,
  cf_fn = utilityR::cf.int,
  ...
)
}
\arguments{
\item{.data}{Charging interval data}

\item{time_col}{Interval time column in \code{.data}}

\item{load_col}{Column in \code{.data} with charging load}

\item{agg_fn}{Function to use to aggregate load data, default is \code{mean}}

\item{time_agg}{What time interval to aggregate to, default is \code{lubridate::hour}}

\item{cf_fn}{Function to calcualte a confidence interval around the aggregation statistic. Default is \code{utilityR::cf.int}}

\item{...}{Additional arguments to pass to \code{plt_load}}
}
\description{
Plot an aggregated load profile.
}
